+++ OutSouringProject/src/main/java/com/sparta/outsouringproject/order/controller/OrderController.java
@@ -17,7 +17,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api/orders")
+@RequestMapping("/api")
 public class OrderController {
 
     private final OrderService orderService;
@@ -25,7 +25,7 @@ public class OrderController {
     /**
      * 주문 요청
      */
-    @PostMapping()
+    @PostMapping("/orders")
     public  ResponseEntity<OrderCreateResponseDto> requestOrder(@RequestBody OrderCreateRequestDto orderRequest){
         return ResponseEntity.status(HttpStatus.CREATED).body(orderService.createOrder(null, orderRequest));
     }
@@ -33,34 +33,34 @@ public class OrderController {
     /**
      * 주문 요청 수락
      */
-    @PatchMapping("/{orderId}/accept")
-    public  ResponseEntity<Void> acceptOrder(@PathVariable("orderId") Long orderId){
-        orderService.acceptOrder(null, orderId);
+    @PatchMapping("stores/{storeId}/orders/{orderId}/accept")
+    public  ResponseEntity<Void> acceptOrder(@PathVariable("storeId") Long storeId, @PathVariable("orderId") Long orderId){
+        orderService.acceptOrder(null, storeId, orderId);
         return ResponseEntity.ok().build();
     }
 
     /**
      * 주문 배달 출발
      */
-    @PatchMapping("/{orderId}/delivery")
-    public  ResponseEntity<Void> startDelivery(@PathVariable("orderId") Long orderId) {
-        orderService.startDelivery(null, orderId);
+    @PatchMapping("stores/{storeId}/orders/{orderId}/delivery")
+    public  ResponseEntity<Void> startDelivery(@PathVariable("storeId") Long storeId, @PathVariable("orderId") Long orderId) {
+        orderService.startDelivery(null, storeId, orderId);
         return ResponseEntity.ok().build();
     }
 
     /**
      * 배달완료
      */
-    @PatchMapping("/{orderId}/complete")
-    public ResponseEntity<Void> completeDelivery(@PathVariable("orderId") Long orderId) {
-        orderService.completeOrder(null, orderId);
+    @PatchMapping("stores/{storeId}/orders/{orderId}/complete")
+    public ResponseEntity<Void> completeDelivery(@PathVariable("storeId") Long storeId, @PathVariable("orderId") Long orderId) {
+        orderService.completeOrder(null, storeId, orderId);
         return ResponseEntity.ok().build();
     }
 
     /**
      * 현재 주문 상태
      */
-    @GetMapping("/{orderId}")
+    @GetMapping("orders/{orderId}")
     public ResponseEntity<OrderStatusResponseDto> currentOrderStatus(@PathVariable("orderId") Long orderId) {
         return new ResponseEntity<>(orderService.getCurrentOrderStatus(null, orderId), HttpStatus.OK);
     }
